Добавлены исправленные классы
