+++ ConsoleApp4/ClassLibrary1/HRDepartment.Domain.csproj
