--- HRDepartment/HRDepartment.Tests/HRDepartment.Tests.csproj
@@ -1,14 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\ConsoleApp2\ConsoleApp2.csproj" />
-    <ProjectReference Include="..\HRDepartment.Domain\HRDepartment.Domain.csproj" />
-  </ItemGroup>
-
-</Project>
