--- HRDepartment/HRDepartment.Tests/HRDepartmentTest.cs
@@ -1,1 +1,0 @@
-﻿
